requirements.txt
streamlit
google-generativeai
pillow
replicate